Add helpers for finding and reserving candidates
The strategy here is to use `pg_advisory_lock` to implement acquiring
and releasing the mutexes while jobs are being processed.

We can do a bit better here, a consequence of the current implementation
is that a consumer will naively fetch a batch of jobs, and attempt to
acquire a held mutex. Better would be to exclude any mutexes that are
currently reserved from the candidate filter.