Add option to delete other jobs on conclusion
For deduplicating/debouncing/throttling purposes, it is sometimes useful
to only handle single instances of jobs.

A typical example would be a job that filled caches or did indexing,
where if we already did the job at timestamp `T`, all identical jobs
that were scheduled prior to `T` or now obsolete.